Add methods and logic to the program

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -4,31 +4,52 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Resources\AttendeeResource;
+use App\http\traits\CanLoadRelationships;
 use App\Models\Attendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
 
 class AttendeeController extends Controller
 {
+    // APPLY THE MIDDLEWARE HERE
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum')->except(['index', 'show', 'update']);
+    }
+
+    use CanLoadRelationships;
     /**
      * Display a listing of the resource.
      */
+    private array $relations = ['user', 'event', 'event.attendees'];
     public function index(Event $event)
     { 
         // Return attendees
-        $attendees = $event->attendees()->latest();
-
-        return AttendeeResource::collection($attendees->paginate());
+        $query = $this->loadRelationship($event->attendees());
+        return AttendeeResource::collection($query->paginate());
     }
+
+    protected function shouldIncludeRelation(string $relation) : bool 
+    {
+        $include = request()->query('include');
+
+        if(!$include){
+            return false;
+        }
+
+        $relations = array_map('trim', explode(',', $include));
+        return in_array($relation, $relations);
+    }
+
 
     /**
      * Store a newly created resource in storage.
      */
     public function store(Request $request, Event $event)
     {
-        $attendee = $event->attendees()->create([
-            'user_id' => 1
-        ]);
+        $attendee = $this->loadRelationship($event->attendees()->create([
+            'user_id' => $request->user()->id
+        ]));
 
         return new AttendeeResource($attendee);
     }
@@ -38,14 +59,15 @@ class AttendeeController extends Controller
      */
     public function show(Event $event, Attendee $attendee)
     {
-        return new AttendeeResource($attendee);
+        return new AttendeeResource($this->loadRelationship($attendee));
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id, Attendee $attendee)
+    public function destroy(Event $event, Attendee $attendee)
     {
+        $this->authorize('delete-attendee', [$event, $attendee]);
         $attendee->delete();
 
         return response(status: 204);
